requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
transformers
torch
torchaudio 
python-multipart
ffmpeg
